+++ tests/system-integration-test/sit-jee-stateless-consumer/src/main/java/io/joynr/systemintegrationtest/jeestatelessasync/SitStatelessAsyncConsumerControllerBean.java
@@ -95,7 +95,7 @@ public class SitStatelessAsyncConsumerControllerBean implements SitControllerSyn
             SystemIntegrationTestStatelessAsync proxy = serviceLocator.builder(SystemIntegrationTestStatelessAsync.class,
                                                                                domain)
                                                                       .withUseCase(USE_CASE)
-                                                                      .withDicoveryQos(discoveryQos)
+                                                                      .withDiscoveryQos(discoveryQos)
                                                                       .withMessagingQos(new MessagingQos())
                                                                       .build();